+++ packages/client/src/registry/StreamRegistry.ts
@@ -3,7 +3,7 @@ import type { StreamRegistryV4 as StreamRegistryContract } from '../ethereumArti
 import StreamRegistryArtifact from '../ethereumArtifacts/StreamRegistryV4Abi.json'
 import { BigNumber } from '@ethersproject/bignumber'
 import { Provider } from '@ethersproject/providers'
-import { scoped, Lifecycle, inject, delay } from 'tsyringe'
+import { delay, inject, Lifecycle, scoped } from 'tsyringe'
 import { getStreamRegistryChainProviders, getStreamRegistryOverrides } from '../Ethereum'
 import { until } from '../utils/promises'
 import { ConfigInjectionToken, StrictStreamrClientConfig } from '../Config'
@@ -12,11 +12,11 @@ import { NotFoundError } from '../HttpUtil'
 import { StreamID, StreamIDUtils } from '@streamr/protocol'
 import { StreamIDBuilder } from '../StreamIDBuilder'
 import { SynchronizedGraphQLClient } from '../utils/SynchronizedGraphQLClient'
-import { searchStreams as _searchStreams, SearchStreamsPermissionFilter } from './searchStreams'
+import { searchStreams as _searchStreams, SearchStreamsPermissionFilter, SearchStreamsOrderBy } from './searchStreams'
 import { filter, map } from '../utils/GeneratorUtils'
-import { ObservableContract, waitForTx, queryAllReadonlyContracts } from '../utils/contract'
+import { ObservableContract, queryAllReadonlyContracts, waitForTx } from '../utils/contract'
 import {
-    StreamPermission,
+    ChainPermissions,
     convertChainPermissionsToStreamPermissions,
     convertStreamPermissionsToChainPermission,
     isPublicPermissionAssignment,
@@ -25,8 +25,8 @@ import {
     PermissionQuery,
     PermissionQueryResult,
     PUBLIC_PERMISSION_ADDRESS,
-    streamPermissionToSolidityType,
-    ChainPermissions
+    StreamPermission,
+    streamPermissionToSolidityType
 } from '../permission'
 import { StreamRegistryCached } from './StreamRegistryCached'
 import { Authentication, AuthenticationInjectionToken } from '../Authentication'
@@ -209,10 +209,15 @@ export class StreamRegistry {
         return this.parseStream(streamId, metadata)
     }
 
-    searchStreams(term: string | undefined, permissionFilter: SearchStreamsPermissionFilter | undefined): AsyncIterable<Stream> {
+    searchStreams(
+        term: string | undefined,
+        permissionFilter: SearchStreamsPermissionFilter | undefined,
+        orderBy: SearchStreamsOrderBy
+    ): AsyncIterable<Stream> {
         return _searchStreams(
             term,
             permissionFilter,
+            orderBy,
             this.graphQLClient,
             (id: StreamID, metadata: string) => this.parseStream(id, metadata),
             this.logger)