+++ packages/client/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add boolean option `raw` to method `.subscribe` to allow raw subscription a stream
+- Add boolean option `raw` to method `.subscribe` to allow raw subscription to a stream
 
 ### Changed