+++ packages/client/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add boolean option `raw` to method `.subscribe` to allow raw subscription to a stream
+- Add optional parameter to `.searchStreams` for defining order of the result set
 
 ### Changed