+++ packages/client/CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fix crashing issue in network library when `acceptProxyConnections` is enabled.
+
 ### Security